Rename test helpers
