+++ internal/testutils/cmd.go
@@ -1,4 +1,4 @@
-package testHelpers
+package testutils
 
 import (
 	"testing"
